fix: uniform navbar and remove flickering
