+++ apps/frontend/src/components/navbar/Navbar.tsx
@@ -1,47 +1,15 @@
-import { Box, BoxProps, defineStyle, Flex, FlexProps, Link } from "@chakra-ui/react";
-import { ConnectButton } from "../ui/ConnectButton";
-import { VotingPowerModal } from "../proposal/VotingPowerModal";
+import { VotingPowerModal } from "@/components/proposal/VotingPowerModal";
+import { ConnectButton } from "@/components/ui/ConnectButton";
+import { VoteLogo } from "@/components/ui/VoteLogo";
+import { useUser } from "@/contexts/UserProvider";
+import { Box, Container, defineStyle, Flex } from "@chakra-ui/react";
 import { useWallet } from "@vechain/vechain-kit";
-import { VoteLogo } from "../ui/VoteLogo";
-
-export const NavbarContainer = ({ children, ...restProps }: BoxProps) => {
-  return (
-    <Box
-      transition={"all 0.3s"}
-      paddingX={{ base: 6, md: 20 }}
-      position={"fixed"}
-      top={0}
-      left={0}
-      right={0}
-      zIndex={100}
-      {...restProps}>
-      {children}
-    </Box>
-  );
-};
-
-export const NavbarInnerContainer = ({ children, ...restProps }: FlexProps) => {
-  const { connection } = useWallet();
-  return (
-    <Flex
-      transition={"all 0.3s"}
-      borderRadius={12}
-      maxWidth={"1440px"}
-      marginX={"auto"}
-      justifyContent={"space-between"}
-      alignItems={"center"}
-      gap={6}
-      {...restProps}>
-      {children}
-      {!connection.isLoading && (
-        <Flex alignItems={"center"} gap={{ base: 3, md: 6 }}>
-          {connection.isConnected && <VotingPowerModal />}
-          <ConnectButton bg={"primary.700"} />
-        </Flex>
-      )}
-    </Flex>
-  );
-};
+import { Link } from "react-router-dom";
+import { useMemo } from "react";
+import { useProposal } from "@/components/proposal/ProposalProvider";
+import { DeleteEditProposal } from "./components/DeleteEditProposal";
+import { CancelProposal } from "./components/CancelProposal";
+import { ExecuteModal } from "./components/ExecuteModal";
 
 export const bgHeaderStyle = defineStyle({
   bgImage: { base: "/images/banner-bg-mobile.png", md: "/images/banner-bg.png" },
@@ -53,13 +21,47 @@ export const bgHeaderStyle = defineStyle({
 });
 
 export const Navbar = () => {
+  const { connection } = useWallet();
   return (
-    <NavbarContainer {...bgHeaderStyle}>
-      <NavbarInnerContainer paddingY={4}>
-        <Link href="/" display="flex" alignItems="center">
-          <VoteLogo width={{ base: 16, md: 20 }} />
-        </Link>
-      </NavbarInnerContainer>
-    </NavbarContainer>
+    <Box {...bgHeaderStyle} position="fixed" top="0" width="100%" zIndex={1000}>
+      <Container maxWidth={"1440px"} marginX={"auto"}>
+        <Flex justifyContent={"space-between"} alignItems={"center"} gap={6} paddingY={4} paddingX={{ base: 4, md: 4 }}>
+          <Link to="/">
+            <VoteLogo width={{ base: 16, md: 20 }} />
+          </Link>
+
+          <Flex alignItems={"center"} gap={{ base: 3, md: 6 }}>
+            {connection.isConnected && !connection.isLoading && <ProposalNavbarActions />}
+            {connection.isConnected && !connection.isLoading && <VotingPowerModal />}
+            <ConnectButton />
+          </Flex>
+        </Flex>
+      </Container>
+    </Box>
+  );
+};
+
+const ProposalNavbarActions = () => {
+  const { proposal } = useProposal();
+  const { isExecutor, isWhitelisted } = useUser();
+
+  const canCancel = useMemo(
+    () => isWhitelisted && ["upcoming"].includes(proposal?.status || ""),
+    [isWhitelisted, proposal?.status],
+  );
+
+  const canEditDraft = useMemo(
+    () => isWhitelisted && ["draft"].includes(proposal?.status || ""),
+    [isWhitelisted, proposal?.status],
+  );
+
+  if (!proposal || proposal.id === "default") return null;
+
+  return (
+    <Flex alignItems={"center"} gap={2} marginLeft={"auto"}>
+      {canEditDraft && <DeleteEditProposal />}
+      {canCancel && <CancelProposal proposalId={proposal?.id} />}
+      {isExecutor && proposal?.status === "approved" && <ExecuteModal proposalId={proposal?.id} />}
+    </Flex>
   );
 };
