+++ apps/frontend/src/components/proposal/ProposalProvider.tsx
@@ -2,28 +2,30 @@ import { ProposalCardType, ProposalStatus } from "@/types/proposal";
 import { Delta } from "quill";
 import { createContext, PropsWithChildren, useContext, useMemo } from "react";
 
-export const ProposalContext = createContext<{ proposal: ProposalCardType }>({
-  proposal: {
-    id: "1",
-    proposer: "0x",
-    createdAt: new Date(),
-    headerImage: {
-      type: "image/png",
-      name: "mock image",
-      size: 1,
-      url: "",
-    },
-    status: ProposalStatus.UPCOMING,
-    description: new Delta().ops,
-    title: "",
-    startDate: new Date(),
-    endDate: new Date(),
-    votingQuestion: "",
+const DEFAULT_PROPOSAL = {
+  id: "default",
+  proposer: "0x",
+  createdAt: new Date(),
+  headerImage: {
+    type: "image/png",
+    name: "mock image",
+    size: 1,
+    url: "",
   },
+  status: ProposalStatus.UPCOMING,
+  description: new Delta().ops,
+  title: "",
+  startDate: new Date(),
+  endDate: new Date(),
+  votingQuestion: "",
+};
+
+export const ProposalContext = createContext<{ proposal: ProposalCardType }>({
+  proposal: DEFAULT_PROPOSAL,
 });
 
-export const ProposalProvider = ({ children, proposal }: PropsWithChildren<{ proposal: ProposalCardType }>) => {
-  const value = useMemo(() => ({ proposal }), [proposal]);
+export const ProposalProvider = ({ children, proposal }: PropsWithChildren<{ proposal?: ProposalCardType }>) => {
+  const value = useMemo(() => ({ proposal: proposal || DEFAULT_PROPOSAL }), [proposal]);
   return <ProposalContext.Provider value={value}>{children}</ProposalContext.Provider>;
 };
