+++ apps/frontend/src/components/ui/ConnectButton.tsx
@@ -2,15 +2,26 @@ import { useI18nContext } from "@/i18n/i18n-react";
 import { WalletIcon } from "@/icons";
 import { Avatar, Button, ButtonProps, Flex, Icon, Text } from "@chakra-ui/react";
 import { useAccountModal, useConnectModal, useWallet } from "@vechain/vechain-kit";
+import { useEffect, useState } from "react";
 
 export const ConnectButton = ({ text, ...props }: ButtonProps & { text?: string }) => {
   const { connection, account } = useWallet();
+  const [canRender, setCanRender] = useState(false);
 
   const { LL } = useI18nContext();
   const { open: openConnectModal } = useConnectModal();
   const { open: openAccountModal } = useAccountModal();
 
-  if (connection.isLoading) return <Flex height={12} />;
+  // add a rendering debouncer to avoid flickering
+  useEffect(() => {
+    if (!connection.isLoading) {
+      setTimeout(() => {
+        setCanRender(true);
+      }, 100);
+    }
+  }, [connection.isLoading]);
+
+  if (connection.isLoading || !canRender) return <Flex height={12} width={8} />;
   if (!connection.isConnected)
     return (
       <StyledButton leftIcon={<Icon as={WalletIcon} />} onClick={openConnectModal} {...props}>
