+++ apps/frontend/src/pages/Proposal/Proposal.tsx
@@ -12,12 +12,12 @@ import { Routes } from "@/types/routes";
 import { BuyNodeCta } from "./components/BuyNodeCta";
 import { ProposalHeader } from "./components/ProposalHeader";
 import { DescriptionSection } from "./components/DescriptionSection";
-import { ProposalNavbar } from "./components/ProposalNavbar/ProposalNavbar";
 import { VotingAndTimeline } from "./components/VotingAndTimeline/VotingAndTimeline";
 import { ProposalStatus } from "@/types/proposal";
 import { CanceledProposal } from "./components/CanceledProposal";
+import { Navbar } from "@/components/navbar/Navbar";
 
-export const ProposalContent = () => {
+export const Proposal = () => {
   const { LL } = useI18nContext();
   const { draftProposal } = useCreateProposal();
   const { account } = useWallet();
@@ -44,50 +44,53 @@ export const ProposalContent = () => {
   }, [account?.address, navigate, params.proposalId]);
 
   if (isLoading) {
-    return <SingleProposalSkeleton />;
+    return (
+      <ProposalProvider proposal={proposal}>
+        <Box bg={"white"}>
+          <Navbar />
+          <PageContainer bg={"white"} pt={{ base: 24, md: 32 }} pb={10}>
+            <SingleProposalSkeleton />
+          </PageContainer>
+        </Box>
+      </ProposalProvider>
+    );
   }
 
-  if (!proposal) {
+  if (!proposal || proposal.id === "default") {
     navigate(`${Routes.HOME}?proposalNotFound=true`);
     return null;
   }
 
   return (
     <ProposalProvider proposal={proposal}>
-      <VStack gap={10} w={"full"} alignItems={"stretch"}>
-        <Flex gap={1} alignItems={"center"} fontSize={"14px"} fontWeight={500}>
-          <Text color={"gray.600"} onClick={() => navigate(Routes.HOME)} cursor={"pointer"}>
-            {LL.homepage()}
-          </Text>
-          <Text color={"gray.400"}>{"→"}</Text>
-          <Text color={"gray.600"}>{LL.proposal.title()}</Text>
-        </Flex>
-        <Stack direction={{ base: "column", md: "row" }} w={"full"} gap={{ base: 10, md: 12 }}>
-          <VStack gap={10} align="stretch" flex={2}>
-            <ProposalHeader />
-            <Heading fontWeight={500} color={"gray.800"} lineHeight={"1.33"}>
-              {proposal.title}
-            </Heading>
-            {!isMobile && <DescriptionSection />}
+      <Box bg={"white"}>
+        <Navbar />
+        <PageContainer bg={"white"} pt={{ base: 24, md: 32 }} pb={10}>
+          <VStack gap={10} w={"full"} alignItems={"stretch"}>
+            <Flex gap={1} alignItems={"center"} fontSize={"14px"} fontWeight={500}>
+              <Text color={"gray.600"} onClick={() => navigate(Routes.HOME)} cursor={"pointer"}>
+                {LL.homepage()}
+              </Text>
+              <Text color={"gray.400"}>{"→"}</Text>
+              <Text color={"gray.600"}>{LL.proposal.title()}</Text>
+            </Flex>
+            <Stack direction={{ base: "column", md: "row" }} w={"full"} gap={{ base: 10, md: 12 }}>
+              <VStack gap={10} align="stretch" flex={2}>
+                <ProposalHeader />
+                <Heading fontWeight={500} color={"gray.800"} lineHeight={"1.33"}>
+                  {proposal.title}
+                </Heading>
+                {!isMobile && <DescriptionSection />}
+              </VStack>
+              <VStack gap={10} align="stretch" flex={1}>
+                {!isCanceled ? <VotingAndTimeline /> : <CanceledProposal />}
+                {isMobile && <DescriptionSection />}
+              </VStack>
+            </Stack>
+            <BuyNodeCta />
           </VStack>
-          <VStack gap={10} align="stretch" flex={1}>
-            {!isCanceled ? <VotingAndTimeline /> : <CanceledProposal />}
-            {isMobile && <DescriptionSection />}
-          </VStack>
-        </Stack>
-        <BuyNodeCta />
-      </VStack>
+        </PageContainer>
+      </Box>
     </ProposalProvider>
-  );
-};
-
-export const Proposal = () => {
-  return (
-    <Box bg={"white"}>
-      <ProposalNavbar />
-      <PageContainer bg={"white"} pt={{ base: 24, md: 32 }} pb={10}>
-        <ProposalContent />
-      </PageContainer>
-    </Box>
   );
 };
